lob/serializers.py
from .models import Board
from rest_framework import serializers

class BoardSerializer(serializers.ModelSerializer):
    class Meta:
        model = Board
        fields = ('board_title','board_contents','board_writer','board_date')
